+++ duckscript_sdk/src/sdk/std/test/mod.rs
@@ -2,8 +2,7 @@ mod assert;
 mod assert_eq;
 mod assert_error;
 mod assert_fail;
-mod test_file;
-mod test_function;
+mod assert_false;
 
 use crate::utils::pckg;
 use duckscript::types::command::Commands;
@@ -18,8 +17,7 @@ pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptEr
     commands.set(assert_eq::create(&package))?;
     commands.set(assert_error::create(&package))?;
     commands.set(assert_fail::create(&package))?;
-    commands.set(test_file::create(&package))?;
-    commands.set(test_function::create(&package))?;
+    commands.set(assert_false::create(&package))?;
 
     Ok(())
 }